Added a helper function which resets the visual selection ( if any ).
An example where this functionality might be useful would be a date
range picker that has a visible and hidden state. If the user selects
one start date and then the date range picker gets hidden, the consumer
of the package should have a way to reset the visual selection if they
want to.